implemented typescript support
