+++ .gitignore
@@ -34,3 +34,4 @@ yarn-error.log
 package-lock.json
 records.json
 stats.json
+.awcache
